Make sure we call 'mdtool' on it's own first
For some strange reason, if mdtool hasn't been run on it's own
at least once on a machine without any arguments, the pack command
won't succeed.  Presumably it's doing some initialization that
the pack command relies on.

Worse still, if it's not initialized, you'll get a totally misleading
error message indicating that the file to pack doesn't exist, even
though it totally does.